feat(chat): implement message reactions and reply functionality with soft delete support

+++ identity/src/main/java/com/chefkix/identity/dto/response/RecipeCompletionResponse.java
@@ -11,8 +11,8 @@ import lombok.experimental.FieldDefaults;
 public class RecipeCompletionResponse {
   String userId;
   Integer currentLevel;
-  Double currentXP;
-  Double currentXPGoal;
+  Integer currentXP;
+  Integer currentXPGoal;
 
   Long completionCount;
